Corrige CPF para BIGINT e versiona o schema SQL
CPF tem 11 dígitos e não cabia em INT (limite ~2,1 bilhões), então
Pessoa.Cpf passa a ser BIGINT/long, refletido em model, DTOs, DAO,
service e controller. Também versiona o schema.sql (com essa correção
e a de Funcionario.Departamento_Nome nullable) para servir de fonte
de verdade para as próximas entidades.

+++ src/main/java/com/ufrn/imd/cinema/dtos/pessoa/PessoaDtoReq.java
@@ -12,7 +12,7 @@ import java.time.LocalDate;
 public class PessoaDtoReq {
 
     @Positive(message = "CPF deve ser um número positivo")
-    private int cpf;
+    private long cpf;
 
     @NotBlank(message = "Sexo é obrigatório")
     @Pattern(regexp = "[MF]", message = "Sexo deve ser 'M' ou 'F'")
